fix(compliance): schedule the regulatory monitor rescan so the surface self-heals (BI-DA37A602)
The regulatory scan was reachable only through the operator's "Run Scan Now"
button. Once run, it never refreshed — a scan that completed weeks ago kept
painting green "completed · 0 alerts" as a current clean bill of health. The
staleness treatment already shipped (scan-freshness.ts degrades an aged scan to
"stale"); this adds the other half the owner asked for — a scan that actually
refreshes.

- New weekly cron (Mon 06:00 UTC) regulatoryMonitorScanScheduled, plus an
  event-triggered regulatoryMonitorScanRequested for an off-cadence "run now",
  both quiescence-gated and concurrency-limited like the other maintenance jobs.
- Scheduled-jobs catalog gains an editable `regulatory-monitor-scan` entry with a
  runNowEvent, so operators can retune the cadence or run one from the ops surface.
- The scan body moves into a non-action govern runner (runRegulatoryMonitorScan)
  so the queue can invoke it without importing the actions layer (application
  boundary). The "use server" action becomes a thin wrapper that keeps the
  request-only revalidatePath the manual path needs.

## Design grounding

- Existing specs/plans reviewed:
  - docs/superpowers/specs/2026-06-14-data-retention-lifecycle-governance-design.md
    — the data-retention sweep is the precedent for a quiescence-gated maintenance
    cron with a scheduled + requested pair and a catalog runNowEvent; this mirrors
    that shape exactly rather than inventing a new one.
- Current code substrate reviewed:
  - apps/web/lib/queue/functions/data-retention-sweep.ts (function pair template)
  - apps/web/lib/operate/scheduled-jobs/catalog.ts (entry shape, runNowEvent convention)
  - apps/web/lib/actions/regulatory-monitor.ts (triggerRegulatoryMonitorScan already
    accepted "scheduled"; only the schedule was missing)
  - scripts/application-boundaries.json (queue may not import actions → runner in govern)
- Source of truth:
  - The scheduled-jobs catalog is the authoritative classification source; the
    Inngest scheduledFunctions array is the runtime registry. Both updated together.
- Decision:
  - Extend the existing maintenance-cron pattern; no new coordination contract.
    Weekly cadence chosen because regulations change on the order of months and
    36 days was clearly past tolerance.

Docs-Impact-Decision: updated (scheduled-jobs catalog purpose documents the new job for the operator-facing ops surface; no separate doc surface describes this cron)
Process-Spine-Decision: extend-existing (adds a scheduled job mirroring the data-retention-sweep precedent; registered in scheduledFunctions and the authoritative catalog; no new coordination contract or route)
Design-Grounding-Decision: extend-existing (specs/plans reviewed: docs/superpowers/specs/2026-06-14-data-retention-lifecycle-governance-design.md; code substrate: data-retention-sweep.ts + catalog.ts + regulatory-monitor.ts + application-boundaries.json; mirrors the existing maintenance-cron pattern, extracts a govern runner to respect the queue→actions boundary)
